ARC: int vs long
- arch specific bitsperlong.h
- general LP64 related changes
- don't force cast argument to __builtin_arc_sr()
- irqflags: STAT32 is 32-bits but generic raw_local_irq_* expect flags to be long

+++ include/soc/arc/aux.h
@@ -11,7 +11,7 @@
 #define read_aux_reg(r)		__builtin_arc_lr(r)
 
 /* gcc builtin sr needs reg param to be long immediate */
-#define write_aux_reg(r, v)	__builtin_arc_sr((unsigned int)(v), r)
+#define write_aux_reg(r, v)	__builtin_arc_sr(v, r)
 
 #else	/* !CONFIG_ARC */
